+++ src/components/NavBar.js
@@ -1,0 +1,20 @@
+// src/components/NavBar.js
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const NavBar = () => {
+  return (
+    <nav>
+      <ul>
+        <li>
+          <Link to="/">Home</Link>
+        </li>
+        <li>
+          <Link to="/wetwell">WetWell</Link>
+        </li>
+      </ul>
+    </nav>
+  );
+};
+
+export default NavBar;
