+++ src/pages/Home.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Home = () => {
+  return (
+    <div>
+      <h1>Home Page</h1>
+      <p>Welcome to the home page of Design of STPs for Small and Medium Towns!</p>
+    </div>
+  );
+};
+
+export default Home;
